Add Custom Camera functionality; Auto-layout constraints added on 3.5 Storyboard
- TakePicture.h/m modified to have custom camera preview using AVFoundation framework
- ApplyFilters modified to orient images the correct way
- Settings view also modified to orient profile images the correct way
- Small int casting error fixed in Messages.m
- Additional warning messages solved that were caused by storyboard items

+++ Wingman2.0/Messages.m
@@ -90,7 +90,7 @@
             [newMessage setObject:[user objectForKey:@"email"] forKey:@"senderEmail"];
             [newMessage setObject:[match objectForKey:@"email"] forKey:@"receiverEmail"];
             [newMessage setObject:self.currentTextView.text forKey:@"text"];
-            int messageCount = numberOfMessages;
+            int messageCount = (int)numberOfMessages;
             messageCount = messageCount + 1;
             NSNumber *newTextNumber = [NSNumber numberWithInt:messageCount];
             [newMessage setObject:newTextNumber forKey:@"number"];
